Migrate Tls configuration to CassandraSessionConfigurator
